Change (#337): reduced tasks
Reduced the tasks by removing task related
stuff so far are preparation to make the tasks
more generic.
Also moved the functions to create tasks into
the api-endpoints.
