+++ src/Hanami/CMakeLists.txt
@@ -71,7 +71,6 @@ set(SOURCES
     src/api/websocket/file_upload.h
     src/args.h
     src/config.h
-    src/core/cluster/add_tasks.h
     src/core/cluster/cluster.h
     src/core/cluster/cluster_handler.h
     src/core/cluster/cluster_init.h
@@ -177,7 +176,6 @@ set(SOURCES
     src/api/http/v1/user/remove_project_from_user.cpp
     src/api/websocket/cluster_io.cpp
     src/api/websocket/file_upload.cpp
-    src/core/cluster/add_tasks.cpp
     src/core/cluster/cluster.cpp
     src/core/cluster/cluster_handler.cpp
     src/core/cluster/cluster_init.cpp
